telegram_notifications/requirements.txt
python-telegram-bot
python-dotenv
pika
imgkit